feat: enable floating score number when inserting grade

+++ src/main/java/school/hei/haapi/model/Grade.java
@@ -27,7 +27,6 @@ import lombok.ToString;
 @NoArgsConstructor
 @EqualsAndHashCode
 public class Grade implements Serializable {
-  // todo: to review all class
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private String id;
@@ -40,6 +39,6 @@ public class Grade implements Serializable {
   @JoinColumn(name = "exam_id")
   private Exam exam;
 
-  private Integer score;
+  private Double score;
   private Instant creationDatetime;
 }
